Fix: renamed ScenarioList to EditScenarioList
